Add first basic experimental setup a la torchbase
for ImageCLEF's radiology report generation, where the idea is to encode
an image into a vector using simple Conv net and use the embedding to
generate a sequence of tokenized CUI codes. This corresponds to the
first task of the challenge and should later accomodate report
generation into natural language. Since every image can have
varying number of CUI's, either a multi-label multiclass classification
should be used or otherwise CUI's can be considered as a sequence of
tokens. Here, the latter approach is adopted, since it will later
generalize to the task of report generation.